Fixing Add/Edit feedback: Changed hours field, changed content inset

+++ source/ConsultantApp/ConsultantApp.iOS/TimeSheets/TimesheetOverviewViewController/TimesheetOverviewTableViewSource.cs
@@ -57,7 +57,7 @@ namespace ConsultantApp.iOS.TimeEntryViewController
             cell.projectCodeField.Text = curEntry.ProjectCode;
             cell.hoursField.Text = curEntry.Hours.ToString();
 
-				cell.onHoursChanged = ( int newHours) => {
+				cell.onHoursChanged = ( float newHours) => {
                 curEntry.Hours = newHours;
 				};
